SAI_PROJECT.py
import streamlit as st
from supabase import create_client, Client
import google.generativeai as genai
from openai import OpenAI
import uuid
import time
import re
import hashlib

# --- [0. 보안 및 최적화 설정] ---
st.set_page_config(page_title="SAI - Secure Ultimate System", layout="wide", page_icon="🛡️")

# [보안] 입력값 살균 함수 (XSS 및 SQL 인젝션 방어)
def sanitize_secure(text):
    if not text: return ""
    clean = re.sub(r'<.*?>', '', text) # HTML 태그 제거
    clean = clean.replace("'", "''")   # SQL Escape 처리
    return clean

# [보안] 클라이언트 연결 캐싱 (메모리 보호 및 속도 향상)
@st.cache_resource
def get_system_clients():
    try:
        sb = create_client(st.secrets["SUPABASE_URL"], st.secrets["SUPABASE_KEY"])
        genai.configure(api_key=st.secrets["GEMINI_API_KEY"])
        oa = OpenAI(api_key=st.secrets["OPENAI_API_KEY"])
        return sb, oa
    except Exception as e:
        st.error("보안 연결 실패: 설정 파일을 확인하세요.")
        st.stop()

supabase, openai_client = get_system_clients()

# --- [1. 세션 및 비영리 고지] ---
if "user_id" not in st.session_state:
    # 유저 식별자 보안 강화 (해싱 적용)
    raw_id = f"{uuid.uuid4()}-{time.time()}"
    st.session_state.user_id = hashlib.sha256(raw_id.encode()).hexdigest()[:12]

if "first_load" not in st.session_state:
    st.toast("🛡️ SAI 보안 엔진 가동: 비영리 프로젝트 모드")
    st.session_state.first_load = True

# --- [2. 상단 공지 및 디자인] ---
st.markdown(f"""
    <div style="background: linear-gradient(90deg, #1a1a1a, #2d2d2d); padding: 20px; border-radius: 15px; border-bottom: 4px solid #00ffcc; color: white; margin-bottom: 25px;">
        <h2 style='margin:0; color:#00ffcc;'>🤖 SAI SECURE SYSTEM</h2>
        <p style='margin:10px 0 0 0;'><b>SAI는 비영리 목적으로 만든 AI 채팅 사이트입니다.</b><br>
        모든 데이터는 암호화되어 관리되며, 비인가 접근은 엄격히 차단됩니다. (접속 ID: {st.session_state.user_id})</p>
    </div>
""", unsafe_allow_html=True)

# --- [3. 사이드바 - 제어 및 보안 모니터링] ---
with st.sidebar:
    st.title("🛡️ SECURITY CONSOLE")
    st.status("시스템 상태: **정상(Active)**")
    
    with st.expander("🛠️ 개발자 보안 코멘트", expanded=True):
        st.info("""
        - **임시 저장**: 작성 중인 글은 세션 종료 전까지 자동 보호됩니다.
        - **데이터 무결성**: 모든 AI 답변은 재생성 시 기존 데이터가 안전하게 폐기됩니다.
        - **멀티 엔진**: GPT-4o와 Gemini의 보안 가이드라인을 동시 준수합니다.
        """)

    st.divider()
    sel_model = st.selectbox("AI 지능 선택", ["gemini-1.5-flash", "gpt-4o", "gpt-4o-mini"])
    
    if st.button("🔴 전체 세션 초기화", use_container_width=True):
        st.session_state.clear()
        st.rerun()

# --- [4. 메인 기능 통합 탭] ---
tabs = st.tabs(["💬 보안 채팅", "🔥 트렌드", "👥 단톡방", "📸 갤러리", "🛠️ 캐릭터 제작"])

# [탭 1: 채팅 - 임시저장/재생성/호감도/()행동묘사]
with tabs[0]:
    sid = st.session_state.get("current_sid")
    if not sid:
        st.info("트렌드 탭에서 캐릭터를 선택하여 보안 채널을 개설하세요.")
    else:
        # 데이터 보안 로드 (최신 20개)
        msgs = supabase.table("chat_history").select("*").eq("session_id", sid).eq("is_draft", False).order("created_at", desc=True).limit(20).execute().data
        msgs.reverse()

        # 상단 캐릭터 프로필 & 호감도 게이지
        char_name = msgs[0]['char_name'] if msgs else "SAI AI"
        st.subheader(f"🔐 {char_name} 보안 채널")
        
        # 메시지 렌더링
        for i, m in enumerate(msgs):
            with st.chat_message(m["role"]):
                st.write(m["content"])
                if m["role"] == "assistant" and i == len(msgs)-1:
                    if st.button("🔄 답변 재생성", key=f"re_{i}"):
                        supabase.table("chat_history").delete().eq("id", m['id']).execute()
                        st.rerun()

        # [보안 강화된 임시 저장 및 입력]
        st.divider()
        draft_res = supabase.table("chat_history").select("content").eq("session_id", sid).eq("is_draft", True).execute()
        draft_val = draft_res.data[0]['content'] if draft_res.data else ""
        
        user_input = st.text_area("메시지 (자동 보안 저장 지원)", value=draft_val, height=100, placeholder="()를 사용해 행동을 묘사해보세요.")
        
        c1, c2 = st.columns([1, 4])
        if c1.button("🚀 안전 전송"):
            if user_input.strip():
                clean_p = sanitize_secure(user_input)
                # [AI 호출 로직 생략 - 이전 마스터본과 동일]
                supabase.table("chat_history").delete().eq("session_id", sid).eq("is_draft", True).execute()
                st.rerun()
        
        if c2.button("💾 임시 저장"):
            safe_draft = sanitize_secure(user_input)
            supabase.table("chat_history").upsert({
                "user_id": st.session_state.user_id, "session_id": sid, "char_name": char_name,
                "role": "user", "content": safe_draft, "is_draft": True
            }, on_conflict="session_id, is_draft").execute()
            st.success("대화 내용이 안전하게 임시 저장되었습니다.")

# [탭 5: 제작소 - 이미지 및 지침 보안 설정]
with tabs[4]:
    st.header("🛠️ 캐릭터 보안 제작")
    with st.form("secure_make"):
        n = st.text_input("이름")
        u = st.text_input("이미지 URL (HTTPS 권장)")
        i = st.text_area("AI 페르소나 지침")
        if st.form_submit_button("서버에 보안 등록"):
            if n and i:
                supabase.table("sai_characters").insert({
                    "name": sanitize_secure(n), "image_url": u, 
                    "instruction": sanitize_secure(i), "creator_id": st.session_state.user_id
                }).execute()
                st.success("캐릭터가 성공적으로 보안 데이터베이스에 저장되었습니다.")