requirements.txt
streamlit
supabase
google-generativeai
openai